Some minor fixes and base classification transform

+++ aitlas/base/metrics.py
@@ -50,7 +50,7 @@ class RunningScore(object):
     def _fast_hist(self, label_true, label_pred, n_class):
         mask = (label_true >= 0) & (label_true < n_class)
         hist = np.bincount(
-            n_class * label_true[mask].astype(int) + label_pred[mask],
+            n_class * label_true[mask].astype(int) + label_pred[mask].astype(int),
             minlength=n_class ** 2,
         ).reshape(n_class, n_class)
         return hist
